refactor: manager modules: move process and task logic
- Move OS/process-related functions (run_command_detached, kill_process_group,
  is_process_active, etc.) from helpers.rs to core/manager/processes.rs
- Move TaskManager-related logic and spawned task handling from Manager struct
  to core/manager/tasks.rs
- Keep actions.rs focused on idle action logic
- helpers.rs now only contains small utility functions
